requirements.txt
fastapi
fastapi-cli
python-dotenv
uvicorn
pydantic
pymongo
motor
certifi
httpx
mongomock
pytest
pytest-asyncio
pydantic
openai
python-jose[cryptography]
passlib[bcrypt]
python-multipart
websockets
sentence-transformers
scikit-learn